This commit is a port of the latest changes I've made to the Java version, it brings yet another optimization to this DOM implementation. The whole idea is to try and avoid to always create a Text node to hold the value of an attribute. The DOM spec requires it, so we still have to do it in case getFirstChild() is called for instance. The reason attribute values are stored as a list of nodes is so that they can carry more than a simple string. They can also contain EntityReference nodes. However, most of the times people only have a single string that they only set and get through Element.set/getAttribute or Attr.set/getValue. In this new version, the Attr node has a value pointer which can either be the DOMString directly or a pointer to the first ChildNode. A flag tells which one it currently is. Note that while we try to stick with the direct DOMString as much as possible once we've switched to a node there is no going back. This is because we have no way to know whether the application keeps referring to the node we once returned (I guess that's not so true for C++!).
The gain in memory varies on the density of attributes in the document.
But in the tests I've run I've seen up to 12% of memory gain in the Java
version. And the good thing is that it also leads to a slight gain in speed
because we allocate fewer objects! I mean, that's until we have to actually
create the node...

To avoid too much duplicated code, I got rid of ParentNode and renamed
ChildAndParentNode, which I never really liked, to ParentNode for
simplicity, this doesn't make much of a difference in memory usage because
there are only very few objects that are only a Parent. This is only true
now because AttrImpl now inherits directly from NodeImpl and has its own
implementation of the ParentNode's node behavior. So there is still some
duplicated code there.


git-svn-id: https://svn.apache.org/repos/asf/xerces/c/trunk@172503 13f79535-47bb-0310-9956-ffa450edef68