Make some changes in order to better support multi-architecture binaries.
To accomplish this, we need to be able to configure once, and compile multiple
times (once for each architecture). So our checks in configure should, ideally,
be environment-specific, rather than architecture-specific.

In line with this, I modified tests for the integer type selection to look
first for standard types defined by the compiler/environment that give us the
information we need. We first look for (and use, if possible) uint16_t,
for instance, rather than checking whether the size of an integer is 2 bytes.

So as long as the compiler environment knows how to supply the appropriate type,
we don't care that an integer might be 2 bytes on one architecture, and 4 in
another. We can accomodate the build of a mixed architecture binary that might
include both. (Theoretically...)



git-svn-id: https://svn.apache.org/repos/asf/xerces/c/branches/jberry@188798 13f79535-47bb-0310-9956-ffa450edef68